First upload of Simon class
This is the first upload of the Simon class.  I still need to see the full interface of the UI classes so that I can refine how the interface talks to the controller.  As of right now, I've gone with an attribute of the Simon class that contains the list of actions that the user needs to take, and have put it up to the interface to compare the inputs to the desired actions during the user interaction phase.  We may refactor this to the controller after I see the full UI code.