Version .2 of Simon class
I added a bunch of comments to clean up the intent of each of the methods.  I also added a testGameEngine() method that runs through a few of the critical methods to verify functionality of the basics.  Debug information is now displayed in the console only when the isDebug flag is set to true (current default).